fix(numpy): Enable the package to be used when numpy is missing
It seems that I was wrong that just keeping the numpy import out of the SkyMatrix class would result in no effect the Grasshopper plugin or the IronPython environment.

So this commit makes the package functional in cases where numpy is not installed or the environment is IronPython. When using the package in IronPython, the only thing to be aware of is that the slower text-file parsing method will be used in place of the faster numpy one.